fix(websocket): Hoist websocket hook to context provider
